requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=20.1.0=py36h2bbff1b_1
async_generator=1.10=py36h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
ca-certificates=2021.7.5=haa95532_1
certifi=2021.5.30=py36haa95532_0
cffi=1.14.6=py36h2bbff1b_0
click=8.0.1=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=pypi_0
cython=0.29.23=pypi_0
dataclasses=0.8=pyh4f3eec9_6
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py36_0
et-xmlfile=1.1.0=pypi_0
freetype=2.10.4=hd328e21_0
gensim=4.1.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
importlib-metadata=4.8.1=pypi_0
importlib_metadata=4.6.4=hd3eb1b0_0
intel-openmp=2021.3.0=haa95532_3372
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.17.0=py36_0
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py36_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py36haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=py36h2bbff1b_0
matplotlib=3.3.4=pypi_0
mistune=0.8.4=py36he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.0.4=py36h343c172_0
msys2-conda-epoch=20160418=1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py36_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
ninja=1.10.2=h6d14046_1
nltk=3.6.2=pypi_0
notebook=6.4.3=py36haa95532_0
numexpr=2.7.3=pypi_0
numpy=1.19.5=pypi_0
numpy-base=1.17.0=py36hc3f5095_0
olefile=0.46=py36_0
openpyxl=3.0.7=pypi_0
openssl=1.1.1l=h2bbff1b_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.1.5=pypi_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py36haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.2=pypi_0
pip=21.0.1=py36haa95532_0
progressbar=2.5=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyhca03da5_0
prompt_toolkit=3.0.17=hd3eb1b0_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
pyrsistent=0.17.3=py36he774522_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-math=0.0.1=pypi_0
pytorch=1.7.0=py3.6_cuda101_cudnn7_0
pytz=2021.1=pypi_0
pywin32=228=py36hbaba5e8_1
pywinpty=0.5.7=py36_0
pyzmq=22.2.1=py36hd77b12b_1
qt=5.9.7=vc14h73c81de_0
qtconsole=5.1.0=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
regex=2021.8.28=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
smart-open=5.2.1=pypi_0
sqlite=3.36.0=h2bbff1b_0
tables=3.6.1=pypi_0
terminado=0.9.4=py36haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pypi_0
tk=8.6.10=he774522_0
torchaudio=0.7.0=py36
torchvision=0.8.0=py36_cu101
tornado=6.1=py36h2bbff1b_0
tqdm=4.62.2=pypi_0
traitlets=4.3.3=py36_0
typing-extensions=3.10.0.2=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py36_1
wheel=0.37.0=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
